third_party/WebKit/LayoutTests/virtual/threaded/inspector/tracing/timeline-paint/timeline-paint-with-style-recalc-invalidations-expected.txt

Tests the Timeline API instrumentation of paint events with style recalc invalidations.

PASS


Running: testLocalFrame
first paint invalidations[
Exception while running: function () {
        PerformanceTestRunner.dumpInvalidations(
            TimelineModel.TimelineModel.RecordType.Paint, 0, 'first paint invalidations');
        next();
      }
TypeError: Cannot read property 'length' of null
    at Object.InspectorTest.addArray (evaluateInWebInspector0.js:243:31)
    at Object.InspectorTest.dumpInvalidations (evaluateInWebInspector0.js:1341:19)
    at eval (evaluateInWebInspector1.js:65:31)
    at result (evaluateInWebInspector0.js:694:25)
    at Timeline.TimelinePanel.InspectorTest.addSniffer (evaluateInWebInspector0.js:1157:75)
    at Timeline.TimelinePanel.result (evaluateInWebInspector0.js:694:25)
    at Timeline.TimelinePanel.receiver.(anonymous function) (evaluateInWebInspector0.js:724:22)
    at Timeline.TimelineController._finalizeTrace (file:///b/s/w/ir/out/Release/resources/inspector/timeline/timeline_module.js:154:99)
    at setTimeout (file:///b/s/w/ir/out/Release/resources/inspector/timeline/timeline_module.js:153:76)
